Fix a race condition in async mpi affecting limiting executor
The mpi_ring_async_executor_test occasionally failed because
the limiting executor was calling find_if on a vector that
was modified by the mpi polling thread.

Instead of allowing external checks (limiting executor) to
access the vector, we use an atomic counter that only the mpi
polling thread increments/decrements and stores the number
of futures/requests in flight. The limiting executor can safely
read this value from any thread.